feat: split esbuild test
