+++ tests/integration/custom-template/tests/index.test.ts
@@ -3,7 +3,7 @@ import { readFileSync } from 'fs';
 import { modernBuild } from '../../../utils/modernTestUtils';
 
 describe('custom template', () => {
-  it(`should allow to custom template by html.template option`, async () => {
+  test(`should allow to custom template by html.template option`, async () => {
     const appDir = path.resolve(__dirname, '..');
 
     await modernBuild(appDir);
