--- tests/integration/esbuild/test/index.test.ts
@@ -1,44 +1,0 @@
-import path from 'path';
-import { readdirSync, readFileSync, rmdirSync, existsSync } from 'fs';
-import { modernBuild } from '../../../utils/modernTestUtils';
-
-const fixtures = path.resolve(__dirname, '../fixtures');
-
-const getJsFiles = (appDir: string) =>
-  readdirSync(path.resolve(appDir, 'dist/static/js')).filter(filepath =>
-    /\.js$/.test(filepath),
-  );
-
-describe('esbuild', () => {
-  it(`should emitted script files correctly when using esbuild transform`, async () => {
-    const appDir = path.resolve(fixtures, 'transform-and-minify');
-    if (existsSync(path.join(appDir, 'dist'))) {
-      rmdirSync(path.join(appDir, 'dist'), { recursive: true });
-    }
-    await modernBuild(appDir);
-
-    const files = getJsFiles(appDir);
-    const mainFile = files.filter(filepath => filepath.startsWith('main'))[0];
-
-    expect(files.length).toBe(3);
-    expect(
-      readFileSync(path.resolve(appDir, `dist/static/js/${mainFile}`), 'utf8'),
-    ).toContain('children:"helloworld"');
-  });
-
-  it(`should emitted script files correctly when using legacy esbuild minify`, async () => {
-    const appDir = path.resolve(fixtures, 'legacy-minify-js');
-    if (existsSync(path.join(appDir, 'dist'))) {
-      rmdirSync(path.join(appDir, 'dist'), { recursive: true });
-    }
-    await modernBuild(appDir);
-
-    const files = getJsFiles(appDir);
-    const mainFile = files.filter(filepath => filepath.startsWith('main'))[0];
-
-    expect(files.length).toBe(3);
-    expect(
-      readFileSync(path.resolve(appDir, `dist/static/js/${mainFile}`), 'utf8'),
-    ).toContain('children:"helloworld"');
-  });
-});
